Moved logic related tests to logic test suite
